no-mistakes(review): Teach callback-form cache tags; add named endpoint guard
